Removing a flag entry in the cache for one cross cloud unit test (#1848)
## Summary
After [my PR was merged into
common](https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2058),
one particular unit test in CrossCloudGuestAccountTest was failing.
An assertion checks the size of the cache at one point, and the size
expected is 4, while the actual value is 5. 5 is actually expected,
because my PR added in another entry for a sha1-cleared flag.
The next assertion also fails because the key for the flag doesn't
follow the credential id convention. But since the flag isn't associated
with a particular account, it shouldn't have an account id attached to
it.
This PR just removes the flag entry from the map before the assertions
are run. As far as I know, there shouldn't be cases in prod where
assumptions are made based just off of the size of the entire cache (the
methods in SharedPreferencesAccountCredentialCache and SPACCWMC should
only return parts of the cache that are specifically filtered to
accounts, etc.). (but lmk if I've missed something)