Fix exception in SDM  (#1847)
Problem : We got an IcM
(https://portal.microsofticm.com/imp/v3/incidents/details/390194576/home)
in which a single user of yammer app reported that the app was crashing
upon launch in SDM.
Check the IcM sumamry for the stacktrace. The issue is that we are
getting a null localAccountId from the cache. We could not figure out
why this issue is happening on 1 particular device as we did not get
more logs requested.

Fix :  To avoid NPE, we simply added a null check

Note : This does not fix the original problem of localAccountId being
NULL but at least can help fix the app from crashing.