Adding webauthn_capable property option for MSAL config files (#1895)
### Summary
This PR adds logic to handle an MSAL config file property
`webauthn_capable` with a boolean value.
When the property is set to true, MSAL adds the `webauthn=1` parameter
to the extra query string parameter list for interactive token calls.
Once the server side implementation handling the parameter is in prod,
adding this extra query parameter will make a passkey option appear on
the sign in page in browser and WebView.
The property is set to false by default, as host apps (only 1P apps)
should set this to true once they can confirm the future passkey feature
works with their app.

### Testing
I used msalTestApp and tested with a test flight of ESTS and some
injected JavaScript code (which mimics the work that still needs to be
done on the ESTS UX side). I set the config property to "true" in the
WEBVIEW test config file, and when using this config file, the passkey
option correctly appeared. The option does not appear when this config
property isn't set.
Note that this PR does not include the logic for the test flight nor the
JavaScript code, so this property will add the query string parameter,
but the server side does nothing about it at the moment.

### Related PRs
-
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2171
- https://github.com/AzureAD/ad-accounts-for-android/pull/2527