Moving UI tests for NAA to msalautomationapp from brokerautomationapp (#1911)
Moving UI tests for NAA to msalautomationapp from brokerautomationapp
for below reasons

- The tests added are for OneAuth and we do not have a OneAuth specific
automation app. We usually have client specific tests in
msalautomationapp. So I think these tests fit in msalautomationapp
- We need @LocalBrokerHostDebugUiTest on all these tests as we need to
specifically run these tests with brokerHost app (to be able to enable
PRTV3 flight dynamically). Currently, in our pipeline, this annotation
(LocalBrokerHostDebugUiTest) is only supported to run on
msalautomationapp. So, the tests are failing. I have synced with Fadi
and the pipeline can be modified in the future if we have a use case to
run the tests annotated with LocalBrokerHostDebugUiTest in
brokerautomationapp as well. But as of now, we did not take this
approach, because the tests fit better with msalautomationapp.

All the tests added here are from these PRs (I have simply moved the
files)
https://github.com/AzureAD/ad-accounts-for-android/pull/2523/files#diff-3082d89842e3922c2460fcd408fb44f2370f550bf1c433db305905f7e342c074
and https://github.com/AzureAD/ad-accounts-for-android/pull/2555