update TestCase2521768 (#1960)
Why
Now that PRTv3 is rolled out the behavior on the test changed hence this
update.

with PRTV3 the controller factory returns the BrokerSSOController and
uses the deviceless PRT to update to a device bound PRT silently.

tested using the fadi/nov-ui branch, branch used for the daily UI
testing.

https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1214778&view=results