Add native auth to MSAL test app (#2001)
### For MSAL team:
In order to let the native auth team execute (manual) release tests, an
application is needed that supports the various use flows that are part
of the release tests.

Changes:
- The existing MSAL test app is extended with a Native Auth section,
which is accessed from the side menu.
- Updated CODEOWNERS to reflect native auth ownership of the newly added
classes.

### For native auth team:
Changes compared to previous versions of this application:
- Sign up attributes are no longer hardcoded to `country` and `city`,
but taken dynamically from the UI (through new key and value text
fields).
- AttributesRequired` state will navigate to the new, generic
AttributesFragment. Logs need to be checked for details about which
attributes are required, and thus which keys need to be supplied.
- `CodeFragment` displays API's `channel`, `sentTo` and `codeLength`
fields, which can now be used for validation for correct behaviour.

Note:
This application is not in a state where it can be used as a
CI-generated artefact by the native auth team. Developers using this
application will need to change the config JSON file and update it to
one of several tenant configurations, recompile and run it. Support for
native auth in the Labs environment is needed to make this application
usable as a CI-generated artefact. Once Labs support is added, the
native auth screens will need config selection UI, similar to the
Fragment dedicated to testing `acquireToken()`.