Revert "update nimbus" (#2050)
This reverts #2042 

Why?
This change is breaking CP and blocking the release.
CP cannot update the library right now, because they need to do an IP
scan before updating a lib. That would usually take a couple weeks to
get the results back.
And because the severity part of this vulnerability is unclear, we will
move this change to April release, giving more time to CP folks do their
investigation.