Give BrokerMsalController higher priority than LocalMsalController if available (#2040)
### What 
Giving BrokerMsalController higher priority than LocalMsalController if
available. This means that if Broker is available, msal will first try
to acquire token via broker before trying its localController

### Why
In order to push brokered auth as default when available. And reduce
usage of regular RT, when PRT can be used which is bound to the device.

### How
Changing the order of controllers to put BrokerMsalController (if
eligible) before LocalMsalController

### Testing
Verified AcquireTokenSilent calls reach Broker first.