Modifying NAA tests to run with broker host (#2061)
Issue : NAA Tests in this pipeline run are failing
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1270279&view=ms.vss-test-web.build-test-results-tab&runId=3448652&resultId=100001&paneView=debug
because they can only run with brokerHost app. We have a logic
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/blob/12d96b11ac124211e0cb0c7e0c9e0dadd07199e1/common/src/main/java/com/microsoft/identity/common/internal/platform/AndroidPlatformUtil.java#L140
that checks if the broker app is a debug app that the tests depend on.

Fix : CP and Auth app debug versions are not produced in the pipeline
currently. So we have to rely on brokerHost debug apk produced.