Update Account object in AccountState after access token refresh. (#2085)
When the access token is refreshed the SDK gets also a new id token. The
updated id token needs to be returned to the application. To avoid
changes to the public interface thus having impact to the user of the
SDK this solution updates the Account object in the existing
AccountSate.

---------

Co-authored-by: SammyO <1315628+SammyO@users.noreply.github.com>