Updated Pom task (#2111)
### Summary
Updating the Pom syntax, which will also denote the scope of each
dependency appropriately.
Note the names of the artifacts and that most of the dependencies are
listed under Runtime now instead of Compile. This means that our
external dependencies which we declare with implementation configuration
are no longer leaking to our consumers. The configurations are now
working as described
[here](https://developer.android.com/build/dependencies#dependency_configurations).