Commenting out the extra MSAL renaming configuration (#2130)
### Summary
After I made some changes related to the POM logic, we see that this
renaming task config (which was originally added as a workaround) is
renaming the aar file earlier than expected for the vsts publish task,
which is still looking for a file named msal-dist-release.aar. On the
other hand, this renaming task config is necessary for Maven publishing.
So while I'm investigating (wondering why this is an issue for only MSAL
and not common), this PR comments out the logic in order to unblock the
daily pipeline. But it should be uncommented while doing the maven
release for MSAL.

---------

Co-authored-by: fadidurah <fadidurah@microsoft.com>