Fix UserAttributesBuilder bug (#2145)
When using UserAttributes.Builder multiple times, e.g. when submitting
attributes across multiple steps/screens, previously set data is
persisted. This is due to the Builder (unintentionally) being a
singleton. The impact of this is that unexpected/unintended user
attributes are sent to the server, causing an incorrect user flow.

This PR extract related changes from
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/2124/commits/0c8665134a8daa58526fc9a77832b8bac3067b7c