Authority data is persisted across different SDK instances  (#2159)
### The target of this branch is dev

The working history please refer this closed PR
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/2151

Company PR:
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2462

### Investigation Summary:
Please refer company PR for root cause description. In this PR, we apply
the new overload method in Authority to pass in the client id. Authority
returns CIAM Authority when client id is null. In the later validation
step, it will be converted into Native Authority.