Move native auth interface towards parameters class methods, Fixes AB#3117217 (#2245)
Fixes
[AB#3117217](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3117217)

Add parameters class methods to native auth interface. Deprecate
parameter list methods.
The signIn with continuation token method has a custom name for Java to
avoid ambiguity. Without this custom name the introduction of the new
method could introduce a breaking change in case the external developer
used null as scopes value.

#### Additional info
The ideal test strategy for these new methods would be wrapping the
`CommandDispatcher` in a class and inject it a spy class during tests to
check that the new parameters class is used properly. This refactor
would require quite some time, so this work will be done in a separate
PBI.