Add null checks for guest account ids, Fixes AB#3330213 (#2361)
### Summary

[AB#3330213](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3330213)

We got a report from a partner team that they were seeing very low
amounts (like 5 instances) of an NPE related to AccountAdapter. Looking
at past related issues, this seems like some account entry data got
malformed for particular accounts/devices, perhaps due to very old
broker versions. The patch in this PR is to add null checks for both the
home and local id from the guest account, and if either are null, then
it won't be added to the guest account array.