Move our Libraries to only use artifact feed for packages, no Maven Central, Fixes AB#3395409 (#2413)
Received an s360 item asking our libraries to always pull packages from
feeds, and not declare mavenCentral as a repository to source
dependencies. The feed itself has maven as an upstream. We cannot use
AndroidADAL feed as that one has some versions deleted (once a version
is deleted from a feed, it cannot be restored, even if an upstream has
it). Created a new feed to support this
https://identitydivision.visualstudio.com/Engineering/_artifacts/feed/NewAndroid.
We pull packages from this new feed, but we still publish our versions
to AndroidADAL. The new feed has AndroidADAL as an upstream still, so
our existing collection of artifacts is still accessible.

I also took this chance to consolidate the Maven VSTS Username and
Access Token fields, which are fields used to authenticate and pull
artifacts from the feed. Previously, each library had their own names
for these fields with an identifier based on the library in question.
These PRs make it so all libraries use the same names,
`ENV_VSTS_MVN_CRED_USERNAME` and `ENV_VSTS_MVN_CRED_ACCESSTOKEN`.

Validation:
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1568877&view=results


[AB#3395409](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3395409)